+++ lib/web_translate_it/configuration.rb
@@ -9,9 +9,9 @@ module WebTranslateIt
   class Configuration
     require 'yaml'
     require 'fileutils'
-    attr_accessor :path, :api_key, :files, :ignore_locales, :logger
+    attr_accessor :path, :api_key, :source_locale, :target_locales, :files, :ignore_locales, :logger
     
-    # Load the configuration file from the path.
+    # Load configuration file from the path.
     def initialize(root_path=RAILS_ROOT, path_to_config="config/translation.yml")
       self.path           = root_path
       configuration       = YAML.load_file(File.join(root_path, path_to_config))
@@ -22,6 +22,7 @@ module WebTranslateIt
       configuration['files'].each do |file_id, file_path|
         self.files.push(TranslationFile.new(file_id, root_path + '/' + file_path, api_key))
       end
+      set_locales
     end
     
     # Makes an API request to fetch the list of the different locales for a project.
@@ -29,25 +30,16 @@ module WebTranslateIt
     #
     #   configuration = WebTranslateIt::Configuration.new
     #   locales = configuration.locales # returns an array of locales: ['en', 'fr', 'es', ...]
-    #
-    # TODO: Make this use the new endpoint serving YAML
-    def locales
-      WebTranslateIt::Util.http_connection do |http|
-        request  = Net::HTTP::Get.new(api_url)
-        response = http.request(request)
-        if response.code.to_i >= 400 and response.code.to_i < 500
-          puts "----------------------------------------------------------------------"
-          puts "You API key seems to be misconfigured. It is currently #{self.api_key}."
-          puts "Change it in RAILS_ROOT/configuration/translation.yml."
-        else
-          response.body.split
-        end
-      end
+    def set_locales
+      project_info = YAML.load WebTranslateIt::Project.fetch_info(api_key)
+      project = project_info['project']
+      self.source_locale  = project['source_locale']['code']
+      self.target_locales = project['target_locales'].map{|locale| locale['code']}
     end
     
     # Convenience method which returns the endpoint for fetching a list of locales for a project.
     def api_url
-      "/api/projects/#{api_key}/locales"
+      "/api/projects/#{api_key}.yaml"
     end
     
     # Returns a logger. If RAILS_DEFAULT_LOGGER is defined, use it, else, define a new logger.