+++ history.md
@@ -1,5 +1,10 @@
 ## Edge
 
+Warning, some deprecations in this version.
+
+* Remove `wti fetch` and `wti upload`
+* `wti push` now only pushes the master language file. Use `wti push --all` to push all or `wti push -l [locale]` to push a specific locale.
+* `wti pull` now only pulls the target language files. Use `wti pull --all` to pull all or `wti pull -l [locale]` to pull a specific locale.
 * Increase read timeout to 40s — Required for very large projects.
 
 ## Version 1.4.7 /2010-02-05